DSND-2136: Rename child mappers
* Previously named list mappers, these classes no longer map lists
  so have renamed.
