+++ src/main/java/uk/gov/companieshouse/filinghistory/api/mapper/upsert/ChildMapper.java
@@ -2,7 +2,7 @@ package uk.gov.companieshouse.filinghistory.api.mapper.upsert;
 
 import uk.gov.companieshouse.api.filinghistory.InternalFilingHistoryApi;
 
-public interface ChildListMapper<T> {
+public interface ChildMapper<T> {
 
     T mapChild(T child, InternalFilingHistoryApi request);
 }
